Added positioning + updated Readme
Added a function to make all the images appear at a random position. BUG: width and height is not generated yet. Also split up the JS-code base in multiple files to make everything more readable. Finally I updated the instructions in the readme with the new location of the data.